first startings or whatever
